requirements.txt
beautifulsoup4==4.12.2
cleantext==1.1.4
emoji==2.2.0
matplotlib==3.5.3
nltk==3.8.1
numpy==1.21.5
urllib3==1.26.10
pandas<=2.0.0
Pillow<=9.5.0
Pillow<=9.5.0
plotly<=5.14.1
python_dateutil<=2.8.2
Requests<=2.31.0
requests_html<=0.10.0
seaborn<=0.12.2
streamlit<=1.22.0
termcolor<=2.3.0
wordcloud<=1.9.2

pip install requests_html
